requirements.txt
langchain-core
langchain-community
datasets
sentence-transformers
chromadb
streamlit
python-dotenv
torch
transformers
accelerate
tf-keras
sentencepiece